+++ ecmascript/compiler/codegen/maple/maple_be/src/cg/x86_64/x64_emitter.cpp
@@ -191,9 +191,10 @@ Mem X64Emitter::TransferMem(Operand *opnd, uint32 funcUniqueId)
             if (symbolName.size() > 2 && symbolName[2] == '.') {
                 string delimiter = "__";
                 size_t pos = symbolName.find(delimiter);
-                uint32 itsFuncUniqueId =
-                    pos > 3 ? stoi(symbolName.substr(3, pos)) : 0; /* 3: index starts after ".L." */
-                uint32 labelIdx = stoi(symbolName.substr(pos + 2, symbolName.length())); /* 2: delimiter.length() */
+                /* 3: index starts after ".L." */
+                uint32 itsFuncUniqueId = pos > 3 ? static_cast<uint32>(stoi(symbolName.substr(3, pos))) : 0;
+                /* 2: delimiter.length() */
+                uint32 labelIdx = static_cast<uint32_t>(stoi(symbolName.substr(pos + 2, symbolName.length())));
                 symIdx = CalculateLabelSymIdx(itsFuncUniqueId, labelIdx);
             } else {
                 symIdx = symbol->GetNameStrIdx().get();
@@ -1767,7 +1768,8 @@ uint64 X64Emitter::EmitArray(MIRConst &mirConst, CG &cg, bool belongsToDataSec)
             DEBUG_ASSERT(false, "should not run here");
         }
     }
-    int64 iNum = (arrayType.GetSizeArrayItem(0) > 0) ? (static_cast<int64>(arrayType.GetSizeArrayItem(0))) - uNum : 0;
+    int64 iNum = (static_cast<int64>(arrayType.GetSizeArrayItem(0)) > 0) ?
+        (static_cast<int64>(arrayType.GetSizeArrayItem(0))) - static_cast<int64>(uNum) : 0;
     if (iNum > 0) {
         if (uNum > 0) {
             uint64 unInSizeInByte =
